Fixed first failing test.
Set the version of junit dependencies.

Set the intial id of the entity to `null`.
This is used to determine if it is a new entity resulting in an insert as desired or an existing entity which results in an update (of zero rows).

Note: Test does not get picked up by Maven